% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fa_by_id.R
\name{get_fa_by_id}
\alias{get_fa_by_id}
\title{Title get fasta by id from fasta file}
\usage{
get_fa_by_id(geneid, fasta, output)
}
\arguments{
\item{geneid}{a data frame with gene id}

\item{fasta}{a fasta file}

\item{output}{a fasta file}
}
\value{
a fasta file
}
\description{
Title get fasta by id from fasta file
}
\examples{
get_fa_by_id(geneid = './geneid.ID',
fasta = 'Oryza_sativa.pep.fasta',
output = 'tmp3.fasta')

}
